Move stub_core_modules to shared helper
